--- serverless/dynamoplus/http/query.py
@@ -1,8 +1,0 @@
-import os
-import logging
-logging.basicConfig(level=logging.INFO)
-from dynamoplus.http.handler.handler import HttpHandler
-
-def query(event, context):
-    handler = HttpHandler()
-    return handler.query(event['pathParameters'], query_string_parameters=event['queryStringParameters'], body=event["body"], headers=event['headers'])
